Finish moving test graphs into grakn-test-tools
