+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/LazyTest.java
@@ -19,7 +19,7 @@
 package ai.grakn.test.graql.reasoner;
 
 import ai.grakn.GraknGraph;
-import ai.grakn.graphs.GeoGraph;
+import ai.grakn.test.graphs.GeoGraph;
 import ai.grakn.test.graphs.MatrixGraphII;
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.QueryBuilder;
