+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/QueryTest.java
@@ -21,7 +21,7 @@ package ai.grakn.test.graql.reasoner;
 import ai.grakn.GraknGraph;
 import ai.grakn.concept.Concept;
 import ai.grakn.test.graphs.AdmissionsGraph;
-import ai.grakn.graphs.GeoGraph;
+import ai.grakn.test.graphs.GeoGraph;
 import ai.grakn.graql.Graql;
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.admin.Conjunction;
